Correct command parsing for `programs` (#5)
Testing for a particular `#command` argument should be `stringcmp not`,
with a `not`, so currently this identifies every `#command` *other* than
`#mine` as a request for the running user's programs.

Replace this `#command` parsing with the more common `case` style used
in HLM's version (see also watchfor.muf). That lets us also show help
for `#help` and an appropriate error for unknown command arguments.